requirements.txt
# Core
numpy>=1.24.0
pandas>=2.1.0

# Deep Learning
tensorflow==2.20.0
timm>=0.9.2

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Model Training & Evaluation
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.12.2

# MLOps
mlflow>=2.7.0
dvc>=3.4.0
dvc[gdrive]>=3.4.0

# Monitoring
evidently>=0.2.15

# API Deployment
fastapi>=0.103.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Model Export
onnx>=1.15.0
onnxruntime>=1.17.0

# Utils
tqdm>=4.66.0
pyyaml>=6.0
joblib>=1.3.2